Rename _allennlp to allennlp and remove allennlp.py
