--- optuna/integration/allennlp.py
@@ -1,3 +1,0 @@
-from optuna.integration._allennlp.dump_best_config import dump_best_config  # NOQA
-from optuna.integration._allennlp.executor import AllenNLPExecutor  # NOQA
-from optuna.integration._allennlp.pruner import AllenNLPPruningCallback  # NOQA
