+++ optuna/integration/allennlp/__init__.py
@@ -1,0 +1,3 @@
+from optuna.integration.allennlp.dump_best_config import dump_best_config  # NOQA
+from optuna.integration.allennlp.executor import AllenNLPExecutor  # NOQA
+from optuna.integration.allennlp.pruner import AllenNLPPruningCallback  # NOQA
